sysops-toolkit/requirements.txt
flask>=3.0.0
psutil>=5.9.0
python-nmap>=0.7.1
networkx>=3.2.0
matplotlib>=3.8.0
requests>=2.31.0
